Added instruction logic and API endpoint. Made temperature directly settable from the API. More unit and functional tests. Updated README.

+++ chestfreezer-backend/database/db_adapter.py
@@ -12,6 +12,9 @@ import hardware.temperature_probes
 import sqlite3
 import sys
 import time
+import threading
+import control.brew_logic
+from util import misc_utils
 
 cursor = None
 db = None
@@ -19,6 +22,14 @@ db = None
 TEMPERATURE_READINGS_TABLE_NAME = 'temperature_readings'
 PROBES_TABLE_NAME = 'probes'
 INSTRUCTIONS_TABLE_NAME = 'instructions'
+    
+def synchronized(func):
+    """ synchronized method decorator, from http://theorangeduck.com/page/synchronized-python """
+    func.__lock__ = threading.Lock()            
+    def synced_func(*args, **kws):
+        with func.__lock__:
+            return func(*args, **kws)
+    return synced_func
 
 def _is_memory_db():
     """returns true if the database is only in-memory (sqlite3). Otherwise itc can be assumed that MySQL is being used """
@@ -27,18 +38,19 @@ def _is_memory_db():
     elif configuration.db_type() == configuration.DATABASE_IN_MEMORY_CONFIG_VALUE:
         return True
 
-def drop_tables():
-    """ drops all the tables used in the app's db """
+def drop_tables(should_drop_probes_too=True):
+    """ drops (almost) all the tables used in the app's db """
     try:
         cursor.execute("DROP TABLE " + TEMPERATURE_READINGS_TABLE_NAME)
     except:
         pass
-        # its ok, table doesnt exist at all    
-    try:
-        cursor.execute("DROP TABLE " + PROBES_TABLE_NAME)
-    except:
-        pass
-        # likewise
+        # its ok, table doesnt exist at all
+    if should_drop_probes_too:    
+        try:
+            cursor.execute("DROP TABLE " + PROBES_TABLE_NAME)
+        except:
+            pass
+            # likewise
     try:
         cursor.execute("DROP TABLE " + INSTRUCTIONS_TABLE_NAME)
     except:
@@ -54,18 +66,16 @@ def does_table_exist(table_name):
         return cursor.execute("SHOW TABLES LIKE '" + table_name + "'") == 1            
 
 def initialize_tables():
-    """ initializes the 4 required tables """
+    """ initializes the 3 required tables """
     # check if tables exist, init the database
     if not does_table_exist(PROBES_TABLE_NAME):
-        cursor.execute("CREATE TABLE " + PROBES_TABLE_NAME + " (probe_id VARCHAR(12), name VARCHAR(100), master BOOLEAN, PRIMARY KEY(probe_id))");
-        
+        cursor.execute("CREATE TABLE " + PROBES_TABLE_NAME + " (probe_id VARCHAR(12), name VARCHAR(100), master BOOLEAN, PRIMARY KEY(probe_id))");        
     if not does_table_exist(TEMPERATURE_READINGS_TABLE_NAME):    
         sql_statement = "CREATE TABLE " + TEMPERATURE_READINGS_TABLE_NAME + " (probe_id VARCHAR(12) REFERENCES " + PROBES_TABLE_NAME + "(probe_id), temperature_C FLOAT(6,3), timestamp DATETIME)"        
-        cursor.execute(sql_statement);
-    
+        cursor.execute(sql_statement);    
     if not does_table_exist(INSTRUCTIONS_TABLE_NAME):    
         # create temperature readings
-        sql_statement = "CREATE TABLE " + INSTRUCTIONS_TABLE_NAME + " ( temperature_C FLOAT(6,3), start DATETIME, end DATETIME, description TEXT )"        
+        sql_statement = "CREATE TABLE " + INSTRUCTIONS_TABLE_NAME + " (instruction_id INT AUTO_INCREMENT, target_temperature_C FLOAT(6,3), from_timestamp DATETIME, to_timestamp DATETIME, description TEXT, PRIMARY KEY(instruction_id))"        
         cursor.execute(sql_statement);
 
 def connect():    
@@ -85,7 +95,27 @@ def connect():
         drop_tables()
         print 'done.'
     initialize_tables()
-            
+
+@synchronized
+def _store_instruction(instruction):
+    """ stores or updates the given instruction. This is not safe! Do not call this directly, unless you know what you are doing. 
+    Use the brew_logic method that safely stores instructions  """
+    from_timestamp = misc_utils.get_storeable_timestamp(instruction.from_timestamp)    
+    to_timestamp = misc_utils.get_storeable_timestamp(instruction.to_timestamp)
+    cursor.execute("SELECT * FROM " + INSTRUCTIONS_TABLE_NAME + " WHERE instruction_id='" + instruction.instruction_id + "'")
+    results = cursor.fetchall()    
+    if len(results) == 0:
+        if _is_memory_db():
+            insert_sql = "INSERT INTO " + INSTRUCTIONS_TABLE_NAME + " VALUES ('" + str(instruction.instruction_id) + "','" + str(instruction.target_temperature_C) + "','" + from_timestamp + "','" + to_timestamp + "','" + instruction.description + "')"
+        else:
+            insert_sql = "INSERT INTO " + INSTRUCTIONS_TABLE_NAME + " (target_temperature_C,from_timestamp,to_timestamp,description) VALUES ('" + str(instruction.target_temperature_C) + "','" + from_timestamp + "','" + to_timestamp + "','" + instruction.description + "')"
+        cursor.execute(insert_sql)        
+    elif len(results) == 1:        
+        update_sql = "UPDATE " + INSTRUCTIONS_TABLE_NAME + " SET target_temperature_C='" + str(instruction.target_temperature_C) + "',from_timestamp='" + from_timestamp + "',to_timestamp='" + to_timestamp + "',description='" + instruction.description + "'"
+        cursor.execute(update_sql)
+    db.commit()    
+
+@synchronized
 def store_probe(probe, should_overwrite=True):        
     """ stores (with the option to overwrite) a new probe """    
     cursor.execute("SELECT * FROM " + PROBES_TABLE_NAME + " WHERE probe_id='" + probe.probe_id + "'")
@@ -98,22 +128,39 @@ def store_probe(probe, should_overwrite=True):
         if should_overwrite:
             update_sql = "UPDATE " + PROBES_TABLE_NAME + " SET name='" + probe.name + "',master='" + str(int(not probe.master)) + "' WHERE probe_id='" + probe.probe_id + "'"            
             cursor.execute(update_sql)
-            #print 'Updated probe #' + probe.probe_id
+            # print 'Updated probe #' + probe.probe_id
         else:
             print 'Probe #' + probe.probe_id + ' is already registered.'
             return
     db.commit()  
-
+    
+@synchronized
 def store_temperatures(temperature_readings):
     """ stores the given list of temperature readings """
     for temperature_reading in temperature_readings:
         probe_id = temperature_reading.probe_id
-        temperature_C = temperature_reading.temperature_C
-        timestamp = datetime.datetime.fromtimestamp(temperature_reading.timestamp).strftime('%Y-%m-%d %H:%M:%S')
+        temperature_C = temperature_reading.temperature_C        
+        timestamp = misc_utils.get_storeable_timestamp(temperature_reading.timestamp)
         sql_statement = "INSERT INTO " + TEMPERATURE_READINGS_TABLE_NAME + " VALUES ('" + probe_id + "','" + str(temperature_C) + "','" + timestamp + "')"            
         cursor.execute(sql_statement)
     db.commit()    
 
+def _get_datetime(possible_timestamp):
+    """ casts a result from sql to a datetime """
+    try:             
+        timestamp = possible_timestamp.strftime("%s")
+    except:
+        timestamp = int(datetime.datetime.strptime(possible_timestamp, '%Y-%m-%d %H:%M:%S').strftime("%s"))    
+    return timestamp
+
+def get_all_instruction_ids():
+    """ self-explanatory """
+    result = []
+    for instruction in get_all_instructions():
+        result.append(instruction.instruction_id)
+    return result
+
+@synchronized
 def get_temperature_readings(from_timestamp=1, to_timestamp=time.time()):
     """ returns all the temperature readings from/upto the given timestamps """    
     found_temperature_readings = []
@@ -128,10 +175,7 @@ def get_temperature_readings(from_timestamp=1, to_timestamp=time.time()):
     for result in all_results:
         probe_id = result[0]
         temperature_C = result[1]           
-        try:             
-            timestamp = result[2].strftime("%s")
-        except:
-            timestamp = int(datetime.datetime.strptime(result[2], '%Y-%m-%d %H:%M:%S').strftime("%s"))
+        timestamp = _get_datetime(result[2])
         temperature_reading = hardware.temperature_probes.TemperatureReading(probe_id, temperature_C, timestamp)
         # dates must be compared "manually" if we are using sqlite3        
         should_add = (not _is_memory_db()) | (_is_memory_db() & (timestamp >= int(from_timestamp)) & (timestamp <= int(to_timestamp)))
@@ -139,14 +183,34 @@ def get_temperature_readings(from_timestamp=1, to_timestamp=time.time()):
             found_temperature_readings.append(temperature_reading)                
     return found_temperature_readings
 
-def get_instructions_for_time(timestamp):
+def _cursor_row_to_instruction(row):
+    """ gets the list from the given cursor row and creates an instruction object from it """    
+    instruction_id = row[0]
+    target_temperature_C = row[1]    
+    start_timestamp = _get_datetime(row[2])
+    end_timestamp = _get_datetime(row[3])
+    description = row[4]
+    instruction = control.brew_logic.Instruction(instruction_id, target_temperature_C, start_timestamp, end_timestamp, description)
+    return instruction
+
+def _is_time_between(timestamp, from_timestamp, to_timestamp):
+    """ self-explanatory """
+    return (timestamp > from_timestamp) & (timestamp < to_timestamp)
+
+@synchronized
+def get_instructions(from_timestamp=time.time(), to_timestamp=time.time()):
     """ reads the instructions table and returns all the instructions that would be valid for the given time """
     found_instructions = []
-    sql_statement = "SELECT * FROM " + INSTRUCTIONS_TABLE_NAME + " WHERE from_unixtime(" + timestamp + ") BETWEEN from and until"
+    sql_statement = "SELECT * FROM " + INSTRUCTIONS_TABLE_NAME
+    while not control.brew_logic.instruction_thread_in_waiting: time.sleep(0.1)    
     cursor.execute(sql_statement);
     all_results = cursor.fetchall()
     for result in all_results:
-        print result        
+        instruction = _cursor_row_to_instruction(result)
+        one_way_abut = _is_time_between(instruction.from_timestamp, from_timestamp, to_timestamp) | _is_time_between(instruction.to_timestamp, from_timestamp, to_timestamp) 
+        other_way_abut = _is_time_between(from_timestamp, instruction.from_timestamp, instruction.to_timestamp) | _is_time_between(to_timestamp, instruction.from_timestamp, instruction.to_timestamp)
+        if one_way_abut | other_way_abut:
+            found_instructions.append(instruction)                
     return found_instructions
 
 def get_probe_by_id(probe_id):
@@ -155,6 +219,7 @@ def get_probe_by_id(probe_id):
         if probe.probe_id == probe_id:
             return probe
 
+@synchronized
 def get_all_probes():
     """ returns all the probes """
     all_probes = []
@@ -168,5 +233,25 @@ def get_all_probes():
         all_probes.append(probe)
     return all_probes      
 
-   
+@synchronized
+def get_instruction_by_id(instruction_id):
+    for instruction in get_all_instructions():
+        if instruction.instruction_id == instruction_id: return instruction
         
+@synchronized        
+def get_all_instructions():
+    """ returns all the instructions """
+    all_instructions = []
+    cursor.execute("SELECT * FROM " + INSTRUCTIONS_TABLE_NAME)
+    all_results = cursor.fetchall()    
+    for result in all_results:
+        instruction = _cursor_row_to_instruction(result)
+        all_instructions.append(instruction)
+    return all_instructions
+
+@synchronized
+def delete_instruction(instruction_id):
+    """ deletes the instruction that has the given instruction ID """
+    delete_sql = "DELETE FROM " + INSTRUCTIONS_TABLE_NAME + " WHERE instruction_id='%s'" % instruction_id.strip()           
+    cursor.execute(delete_sql)
+    db.commit()
